+++ src/Constants.cs
@@ -4,5 +4,7 @@
     {
         public const string Folder = ".template.config";
         public const string TemplateFileName = "template.json";
+        public const string VsHostFileName = "vs-2017.3.host.json";
+        public const string CliHostFileName = "dotnetcli.host.json";
     }
 }